completed tests for address service

+++ SPARKLES_GARDEN/src/main/java/com/qac/sparkle_gardens/services/AddressService.java
@@ -40,7 +40,9 @@ public class AddressService {
 	 * @param id
 	 */
 	public void deleteAddress(long acctId, long addressId) {
-		
+		if (acctId <= 0 || addressId <= 0) {
+			throw new IllegalArgumentException();
+		}
 		// Retrieve all the addresses under a given address ID
 		ArrayList<CustomerHasAddress> custAdd = (ArrayList<CustomerHasAddress>) custAddressRepository.findByAddressID(addressId);
 		
@@ -70,6 +72,9 @@ public class AddressService {
 	 * @param newAddress
 	 */
 	public void createAddress(Customer customer, Address newAddress) {
+		if (customer == null || newAddress == null) {
+			throw new IllegalArgumentException();
+		}
 		Address address = newAddress;
 		CustomerHasAddress custAdd = new CustomerHasAddress(customer, address); 
 		custAddressRepository.persistCustomerHasAddress(custAdd);
@@ -94,6 +99,10 @@ public class AddressService {
 	 * @param postCode
 	 */
 	public void createAddress(Customer customer, String buildingNum, String addressLine1, String addressLine2, String city, String county, String country, String postCode, String addressType) {
+		if (customer == null || buildingNum == null || addressLine1 == null || addressLine2 == null || city == null || county == null|| country == null|| postCode == null|| addressType == null) {
+			throw new IllegalArgumentException();
+		}
+		
 		Address address = new Address(buildingNum, addressLine1, addressLine2, city, county, country, postCode, addressType);
 		CustomerHasAddress custAdd = new CustomerHasAddress(customer, address); 
 		custAddressRepository.persistCustomerHasAddress(custAdd);
@@ -108,13 +117,45 @@ public class AddressService {
 		
 	}
 	
-	
+	/**
+	 * 
+	 * @param customer
+	 * @param buildingNum
+	 * @param addressLine1
+	 * @param city
+	 * @param county
+	 * @param country
+	 * @param postCode
+	 * @param addressType
+	 */
+	public void createAddress(Customer customer, String buildingNum, String addressLine1, String city, String county, String country, String postCode, String addressType) {
+		if (customer == null || buildingNum == null || addressLine1 == null || city == null || county == null|| country == null|| postCode == null|| addressType == null) {
+			throw new IllegalArgumentException();
+		}
+		
+		Address address = new Address(buildingNum, addressLine1, city, county, country, postCode, addressType);
+		CustomerHasAddress custAdd = new CustomerHasAddress(customer, address); 
+		custAddressRepository.persistCustomerHasAddress(custAdd);
+		
+		/** This code prevents duplicate address objects from being created.
+		 * If the address has not already been created then add a new address 
+		 * to the repository.
+		 */
+		if (!addressRepository.isDuplicate(address)) {
+			addressRepository.persistAddress(address);
+		}
+		
+	}
 	/**
 	 * 
 	 * @param customer
 	 * @param otherAddress
 	 */
 	public void deleteAddress(Customer customer, Address address) {
+		if (customer == null || address == null) {
+			throw new IllegalArgumentException();
+		}
+		
 		CustomerHasAddress custAdd = new CustomerHasAddress(customer, address);
 		custAddressRepository.removeCustomerHasAddress(custAdd);
